requirements.txt
pandas==2.0.3
flask ==2.3.3
Werkzeug==2.3.7
numpy==1.22.3
#numpy
tensorflow == 2.13.0
requests==2.29.0
gunicorn==20.1.0
scikit-learn==1.3.0
rasterio==1.3.8
#tensorflow-cpu==2.4.1
matplotlib==3.7.2
Pillow == 9.2.0
